requirements.txt
nltk
requests
bs4
readability-lxml